+++ CHANGELOG.md
@@ -9,6 +9,20 @@
 
 - Added support for Android 34
 
+- Missing Android microphone permissions will now be gracefully handled.
+
+  When using the JS API, `callInvite.accept()` and `voice.connect()` will now reject the returned Promise with error `31401`.
+
+  When accepting an incoming call through the native notification, the analogous `31401` error can be caught by attaching a listener to `voice.on(Voice.Event.Error, ...)`. See the following example:
+  ```ts
+  voice.on(Voice.Event.Error, (error) => {
+    // handle error
+    if (error.code === 31401) {
+      // show the end-user that they did not give the app the proper permissions
+    }
+  });
+  ```
+
 ## Fixes
 
 ### Platform Specific Fixes
@@ -57,6 +71,14 @@ voice.on(Voice.Event.CallInvite, (callInvite) => {
   | --- | --- |
   | 31210 | Raised when a Call Message is sent with an invalid message type. |
   | 31211 | Raised when a Call Message is sent when the call is not yet ready to send messages. This can typically happen when the Call/CallInvite is not yet in a ringing state. |
+
+### Platform Specific Changes
+
+#### Android
+
+- When permissions are not available, the SDK will now raise a new
+  `PermissionsError` with code `31401` when attempting to make an outgoing call
+  or when trying to accept an incoming call.
 
 1.0.0 (Mar 25, 2024)
 ====================